move mongodb files into mongo folder
